Fix MinGW compiler warning
The compiler warning is:
	field precision specifier '.*' expects argument of type 'int'

So we cast '(c - _pgmptr + 1)' from ptrdiff_t (long long int) to int.

+++ main.c
@@ -858,7 +858,8 @@ static void set_default_vpncscript(void)
 			exit(1);
 		}
 		if (asprintf((char **)&default_vpncscript, "%.*s%s",
-			     (c - _pgmptr + 1), _pgmptr, DEFAULT_VPNCSCRIPT) < 0) {
+			     (int)(c - _pgmptr + 1), _pgmptr,
+			     DEFAULT_VPNCSCRIPT) < 0) {
 			fprintf(stderr, _("Allocation for vpnc-script path failed\n"));
 			exit(1);
 		}
